monolithic app.tsx broken up
